Separate job_flows into two files
