+++ packages/examples/cvat/exchange-oracle/src/cvat/revert_job.py
@@ -1,0 +1,16 @@
+from src.db import SessionLocal
+
+import src.services.cvat as db_service
+import src.cvat.api_calls as cvat_api
+
+
+def revert_job_creation(escrow_address: str) -> None:
+    with SessionLocal.begin() as session:
+        project = db_service.get_project_by_escrow_address(session, escrow_address)
+        if project is not None:
+            if project.cvat_cloudstorage_id:
+                cvat_api.delete_cloustorage(project.cvat_cloudstorage_id)
+            if project.cvat_id:
+                cvat_api.delete_project(project.cvat_id)
+            db_service.delete_project(session, project.id)
+            session.commit()
